Use RecordItem in tests
RecordFileLogger tests are written in a very non-modular way.
Changing storeRecord(..) fans out to changes in 50+ places in tests.
This PR is not changing any prod code (to make review easier), it is
refactoring part of these tests so that the PR changing prod-code looks sane.

Bonus: Since Transfer list is part of TransactionRecord proto, updated
assertRecord() to directly call assertRecordTransfer().

Signed-off-by: Apekshit Sharma <apekshit.sharma@hedera.com>