Update traefik service type (#1844)
We have insecure ports 80 and 5600 and a secure port 443 that uses a self-signed certificate. 
We want to support wildcard certs with DNS over TLS, however Traefik does not support Let's Encrypt with the community edition when there's more than one replica.
As such, as part of our HA story we want to to consolidate the ports so that we can deprecate 5600 and only expose 443 (with 80 redirecting to 443). 

As part of a simplified HA strategy we will move the cert management to the cloud service level, in place of Traefik self-signed cert.
Additionally, we will employ a cloud service loadBalancer in place of Traefik load balancing.

- Update the `README.md` with configurations for a custom values file to the common chart for use when deploying to prod. For example this will contain cloud related configs used when an external load balancer is used instead of Traefik
  - Update Traefik service type to ClusterIP in prod config
  - Update Traefik service with annotations to utilize external load balancer and standalone NEG approach by default - https://cloud.google.com/kubernetes-engine/docs/how-to/standalone-neg
  - Add high level Kubernetes resource overview of traffic flow
  - Remove port 5600 from Traefik exposed ports list
- Increase Netty `maxInboundMetadataSize` from 1kb to 2kb as HTTP/2 headers from load balancer exceed initial limit

Signed-off-by: Nana-EC <nana.essilfie-conduah@hedera.com>