Run transaction via EvmTxProcessor (#3986)
A base branch for a PoC that will proove we can run transaction simulations via EvmTxProcessor that will serve as an entry point. Initially a mocked data will be used.

After the PoC is done, an EVM module encapsulating all of the needed logic, would be extracted. This module will be reused both in hedera-node and mirror-node projects.

Signed-off-by: Ivan Kavaldzhiev <ivankavaldzhiev@Ivans-MacBook-Pro-4.local>