HIP-584: Return 501 unsupported error when calling precompile (#5645)
In order to utilize eth_call usage at its best, we would want to adapt the json-rpc relay to redirect eth_call requests which call contracts, always to the mirror-node. It's possible, however, that in a nested sub-transaction to have a child precompile call. Precompile support is not yet enabled, so we need to have a way to tell the relay about that and then it could redirect the call to the consensus nodes as a fallback.

Signed-off-by: IvanKavaldzhiev <ivankavaldzhiev@gmail.com>