Combine three contract feature scenario outlines into one
While the scenarios ran top to bottom order in IDEA, this is not the case
when running with gradle on the command line. NPEs and chaos resulted.

I did not have any luck solving the scenario ordering problem so I placed
all of the steps within a single scenario to ensure order.

Signed-off-by: Jeff Schmidt <jeffrey.schmidt@swirldslabs.com>