Add Acceptance Test for Hollow Account to Contract (#5688)
Enhance contract acceptance test with the finalizing hollow account to contract scenario as mentioned in HIP-583.

Please refer to contract.feature as its description and the defined steps are pretty explicit on what it took to incorporate the ask of #5423 with the existing contract acceptance test scenario.

This PR modifies several files within hedera-mirror-test/src/test only. Okay and also .gitignore. There are no changes to any other parts of Mirror Node itself.

Signed-off-by: Jeff Schmidt <jeffrey.schmidt@swirldslabs.com>
Signed-off-by: Edwin Greene <edwin@swirldslabs.com>
Co-authored-by: Edwin Greene <edwin@swirldslabs.com>
Signed-off-by: Marc Kriguer <marc.kriguer@swirldslabs.com>