Fix importer crash on startup (#6026)
Prevent lombok generated hashCode() from invoking getNetworkPrefix(). If the networkPrefix property is not set when network is set to OTHER, InvalidConfigurationException will be thrown and while the Spring application context is being created and the downloader will fail to start up.

Rename getNetworkPrefix() to getEffectiveNetwork() so it is not associated with the networkPrefix field.

Signed-off-by: Jeff Schmidt <jeffrey.schmidt@swirldslabs.com>
Signed-off-by: IvanKavaldzhiev <ivankavaldzhiev@gmail.com>