HIP-584: Integrate StackedStateFrames into Updater components (#6066)
We need to integrate the special state for speculative writes into the mirror-node Updater components, so that we keep track of the changes from previous transactions. As an example this might be invoking contractCreate multiple times (several nested contract creates). In addition we need to make possible the Updaters to commit the accumulated changes from a single frame into the state.

Signed-off-by: Ivan Ivanov <ivanivanov.ii726@gmail.com>