HIP-584: Adapt Mirror Archive Node to respect alias addresses (#6454)
Currently when we try to invoke accounts with alias addresses from the in-memory store we don't find them, since we persist them using their long-zero form. This PR fixes the issue and now we first resolve the alias to a long-zero address and then we try to find it from the in-memory store. This issue is present since 0.84.0 release of mirror-node.

Additionally this PR aligns encoding to long zero addresses to use only num values and ignore shard and realm, as is the current logic in hedera-services.

Removed persisting of account from trackLazilyCreatedAccount since we already persisted it in AutoCreationLogic.

Signed-off-by: IvanKavaldzhiev <ivankavaldzhiev@gmail.com>
Signed-off-by: Mustafa Uzun <mustafa.uzun@limechain.tech>
Co-authored-by: Mustafa Uzun <mustafa.uzun@limechain.tech>

(cherry picked from commit 05a09b458744caad066e7e007d3e5140ef11d283)
Signed-off-by: Mustafa Uzun <mustafa.uzun@limechain.tech>