6708 crypto allowance acceptance fails in previewnet (#6711)
A recent change to the allowances REST API is not to return deleted crypto and fungible token allowances. Previously they were returned with zero `amount_granted` and `amount`. 

The crypto and fungible token allowance acceptance tests have been adjusted to reflect this. The change is to expect a 200 response from the API as usual, but then to assert that the allowances list is empty.

---------

Signed-off-by: Jeff Schmidt <jeffrey.schmidt@swirldslabs.com>