HIP-584: Remove InvalidTransactionException dependency from services code (#6767)
As mentioned #6613 (comment) , services code should not have dependencies on web3 code, with the exception of the properties and Store classes. For this reason we need to replace InvalidTransactionException from web3 with InvalidTransactionException from services in the precompiles.

This change in the type of exception also makes HtsPrecompiledContract to not catch the exception properly and the flow goes into the generic catch(Exception e) block. This can cause confusion for the end user because the mirror node sends wrong error message as result.

Signed-off-by: Ivan Ivanov <ivanivanov.ii726@gmail.com>