HIP-584: Change InvalidTransactionException type in Store (#6788)
When the store does not find an object and the OnMissing parameter is set to THROW an InvalidTransactionException is thrown (from the web3 package). In HTSPrecompiledContract#computeInternal we catch InvalidTransactionException from service so in this case the exception is caught in the general Exception block, the frame does not revert and the response is successful although it is not. Junit tests are adapted.

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>