HIP-584: Transfer on estimate gas results in error 400 (#6778)
As a user, I would like to use HTS tokens via ERC20 interface. Gas estimation on the erc20 transfer is not working properly.

Prerequisites:
- token address - `0x0000000000000000000000000000000000000573`
- receiver (account created via hbar transfer) - id `1394` and evm address `0xa0e1d164c8027f8468d2643b714e0637ec2b283b`


PS: Base account creation had to be refactored because there is no test account with a valid evm address. All of the accounts are created with `AccountCreateTransaction` which results in an account with evm address that is just a long zero address of the account id. Creating an account with a valid evm address (not long zero) can be done via `TransferTransaction`.

Return empty precompile result if the sender is missing && isEstimate = true && we have top level token call for transfer, approve, associate and dissociate. These contracts depend on the sender and will revert if this value is missing

Signed-off-by: nikolay <n.atanasow94@gmail.com>
Signed-off-by: Ivan Ivanov <ivanivanov.ii726@gmail.com>
Signed-off-by: Ivan Ivanov <56980328+0xivanov@users.noreply.github.com>
Co-authored-by: Ivan Ivanov <ivanivanov.ii726@gmail.com>
Co-authored-by: Ivan Ivanov <56980328+0xivanov@users.noreply.github.com>