Make `MirrorEntityAccess.isUsable` **not** consider expiry
`MirrorEntityAccess` is using expiry as a gate for `isUsable` for accounts incl contracts

Should be gated on feature flags - but isn't.  Expiry isn't enabled now but this code
thought it was.

Solution for now is to simply _not_ check for expiry as that is the network behavior
right now.

Also, make sure balance behaves properly: >0 balance means isUsable, <0 balance means
_not_ isUsable, and ==0 balance is a special case.

Signed-off-by: David Bakin <117694041+david-bakin-sl@users.noreply.github.com>