Set dynamic value for create token in acceptance tests (#6917)
Exchange rates in local / testnet / previewnet envs are different and token create operations cost 1 usd and 2 usd for custom fees. Currently the value passed in the create token tests are hardcoded and work only in local environment. This PR implements dynamic setting of the hbars sent to the create token call, so the fee is sufficient.

Signed-off-by: Ivan Ivanov <ivanivanov.ii726@gmail.com>