HIP-584: Adapt stack state to utilize pop functionality (#6920)
Currently, the use of the in-memory stacked state frames requires that we add new RWCachingState frame by pushing it on top of the stack for each nested transaction. The normal execution flow would consist of popping this added RWCachingState at the end of each nested transaction during commit(). The current code base does not pop the added frames and the stack keeps increasing in height.
This PR implements the missing pop of the RWCachingStateFrame from the stack at the end of each nested transaction's lifecycle.
More integration tests are added and in all gas estimation tests we are going to check now the height of the stack frame.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>