HIP-584: Estimated gas under expected range (#7041)
When running the acceptance tests in estimate.feature sometimes some of them (such as I call estimateGas with function that changes contract slot information by updating global contract field with the passed argument) fail because the estimated gas is below the lower expected range. After debugging it seems that in some cases the contract's bytecode could not be fetched from the repository and the initial gas estimation is 21000 which is the minimum gas cost and this affects the further gas calculations.

This PR adds check in estimate gas acceptance tests to ensure that the contract is deployed and that the bytecode will be fetched successfully in web3. As a result the gas estimation will be within the expected range.

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>