HIP-584: Adapt stacked state to use ThreadLocals (#6676)
Reducing memory footprint by using ThreadLocals for the state frame caches. K6 tests would be performed to analyze the difference before and after these changes.

The stack inside StackedStateFrames is now wrapped in a ThreadLocal, allowing each transaction (using unique single thread) to have its own stack of frames. A new method is introduces that initialises the stack with the current stackBase (suitable for eth_estimateGas inside the binary search algorithm) or with a brand new clean stackBase. It makes sense to use the cached object in stackBase only for the scope of a single transaction, since it can easily start holding stale data, while there are DB updates at the same time. Thus, stackBase is now wrapped also in ThreadLocal and cleaned after the binary search algorithm is finished.

Since the stack is now a static field (needed by ThreadLocal), we couldn't keep the generic K type for the key, so now the StackedStateFrames are not bound to a specific type but use Object for all key values.

Signed-off-by: IvanKavaldzhiev <ivankavaldzhiev@gmail.com>
Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>
Signed-off-by: Ivan Ivanov <ivanivanov.ii726@gmail.com>
Signed-off-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>
Signed-off-by: nikolay <n.atanasow94@gmail.com>
Signed-off-by: Jeff Schmidt <jeffrey.schmidt@swirldslabs.com>
Signed-off-by: Jesse Nelson <jesse@swirldslabs.com>
Signed-off-by: Xin Li <xin@swirldslabs.com>
Signed-off-by: Steven Sheehy <steven.sheehy@swirldslabs.com>
Signed-off-by: Edwin Greene <edwin@swirldslabs.com>
Signed-off-by: David Bakin <117694041+david-bakin-sl@users.noreply.github.com>
Co-authored-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>
Co-authored-by: Ivan Ivanov <ivanivanov.ii726@gmail.com>
Co-authored-by: Kristiyan Selveliev <138600514+kselveliev@users.noreply.github.com>
Co-authored-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>
Co-authored-by: Ivan Ivanov <56980328+0xivanov@users.noreply.github.com>
Co-authored-by: Nikolay Atanasow <n.atanasow94@gmail.com>
Co-authored-by: Jeff Schmidt <jeffrey.schmidt@swirldslabs.com>
Co-authored-by: Jesse Nelson <jesse@swirldslabs.com>
Co-authored-by: Xin Li <59580070+xin-hedera@users.noreply.github.com>
Co-authored-by: Steven Sheehy <17552371+steven-sheehy@users.noreply.github.com>
Co-authored-by: Edwin Greene <edwin@swirldslabs.com>
Co-authored-by: David Bakin <117694041+david-bakin-sl@users.noreply.github.com>
Co-authored-by: Bilyana Gospodinova <138879266+bilyana-gospodinova@users.noreply.github.com>
Co-authored-by: Steven Sheehy <steven.sheehy@swirldslabs.com>