HIP-584: Fix block hash exception due to hash value not being sliced (#7091)
Fix the IllegalArgumentException ("Hex value is too large: expected at most 32 bytes but got 48") that is thrown when we call the blockhash function in a solidity contract. The exception is thrown in StaticBlockMetaSource#getBlockHash -> org.hyperledger.besu.datatypes.Hash::fromHexString. The cause for this is that the persisted hash values in the record files are of length 48 bytes but org.hyperledger.besu.datatypes.Hash::fromHexString expects the input to be of length 32 bytes. Following HIP-415 (https://hips.hedera.com/hip/hip-415) that was implemented in services the returned hash value from StaticBlockMetaSource#getBlockHash needs to be constructed from the first 32 prefix bytes of the hash.

In this PR some integration tests are added in order to cover some typical scenarios related to HIP-415 such as:

The block hash of the genesis block needs to be all zeroes.
The block hash of the current block needs to be all zeroes.
The block hash of the last 256 blocks (not including the current one) is the 32 byte prefix from the 48 byte hash.
In the integration tests the following changes were made:

The persisted record file hash values are changed to be numeric in order to be parsable later as a hex.
Preserve a record file in ContractCallTestSetup that is not in the latest block in order to get its index (block number) and test scenario 3.

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>