Fix acceptance tests race condition for EstimateFeature (#7125)
Addresses a race condition observed when tests are run concurrently, particularly affecting the handling of tokens and contracts. The race condition could lead to a scenario where a contract and/or token is already created (within another test suite) and loaded from the Map, which leads to networkTransactionResponse being null and throwing java.lang.NullPointerException. To solve this a null check for networkTransactionResponse was added before proceeding to verify the mirror transaction response.

Signed-off-by: Yordan Iliev <yordan.iliev@limechain.tech>