HIP-584: Fix nested contract deploy fail (#7148)
When deploying nested contracts after the Besu upgrade to 23.10.0 (and hedera-evm upgrade to 0.43.0) the target addresses of the created contracts are now created as address aliases based on the sender address and the account nonce. The account nonce was returned always as 0 from the default getNonce(Address) method from HederaEvmEntityAccess and this caused collisions in the generated aliases for the created contracts. This method is now overriden in MirrorEntityAccess to return the correct nonce value from the DB.

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>