HIP-584: Account balance cache delays is reflecting to proper eth_call balance update (#7178)
We've identified an issue where the caching mechanism for an account balance (in HBARS) is causing a delay in reflecting real-time balance updates. When performing an eth_call operation to retrieve an account balance, followed by a transaction to change the account balance, and then another eth_call operation to retrieve the updated balance, the latter eth_call does not return the updated balance immediately. It retains the old balance from the cache for about 30 seconds before updating to reflect the new balance.

Steps to reproduce: 

1. Perform an eth_call operation to an account to retrieve the current balance.
2. Send funds to the account to change its balance and wait for the transaction to be received in the mirror node.
3. Perform another eth_call operation to the account to retrieve the updated balance.

Additional context:
Upon investigation, it appears that the caching mechanism triggered during the first eth_call operation is holding onto the old balance value for a duration of 30 seconds. This delay in cache update is leading to incorrect balance being returned in the subsequent eth_call operation post the account balance change.

Signed-off-by: Ivan Ivanov <ivanivanov.ii726@gmail.com>