HIP-584: 0x returned when contract's runtime bytecode is not available yet (#7203)
Some of the acceptance tests for isToken and getTokenInfo scenarios were failing randomly. The response was HTTP 200 and a body with content "0x". After further investigation it seems that if a contract is not fully deployed (it has an address but the runtime bytecode is not imported yet) then we get this ambiguos response.

In this PR the following changes were made:
- `MirrorEntityAccess#fetchCodeIfPresent` is changed to return null if the entity is not found. This change was necessary because the AbstractCodeCache depends on this value in order to function correctly. Otherwise, the cache would contain addresses as keys and null values for some time even when the code is available at a later point. 
- `MirrorEvmTxProcessorImpl#buildInitialFrame` was changed to correspond to the implementation in services.
- As a result from the other two bullet points in case the runtime bytecode of a contract is not available yet in buildInitialFrame a `MirrorEvmTransactionException` will be thrown and the user will get HTTP 400 with error code "INVALID_TRANSACTION".

The acceptance tests are now enhanced to verify that a contract's runtime bytecode is available and that the token exists before executing the other tests for isToken and getTokenInfo. 

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>