HIP-584: Provide proper estimate on contract deploy (#7327)
Currently when we deploy a contract on consensus node we don't take into account the size of the init bytecode. Mirror node copied this logic and kept the same behaviour.

The json-rpc-relay, however, is taking into account the input size and performs a pre-check on contract deploy. When we try to deploy a contract using the estimate from the mirror node sometimes we hit an issue, since the pre-check on relay side has a higher expected gas value than what comes in as an estimate from the mirror-node.

This PR fixes the gas estimate on mirror-node side by taking into account the init bytecode size. In this way we would get an estimate that is between 5%-20% higher than what the EVM would return as used gas and this value should pass the json-rpc-relay pre-check phase.

Signed-off-by: IvanKavaldzhiev <ivankavaldzhiev@gmail.com>