HIP-584 Historical: Add negative integration tests for historical precompile support - part 2 (#7434)
his PR introduces a change when a block is before the latest but it is not found in the DB. In this case an exception will be thrown instead of a 0x response. This is needed in order to make the behaviour similar to the other cases with error responses.

This PR adds some of the negative integration tests needed to verify extensively the historical support implementation. The tests check that the historical data will not be found if an earlier block is passed. In this case an exception will be thrown.

In a separate PR the tests for GET_CUSTOM_FEES_FOR_TOKEN_WITH_FIXED_FEE, GET_CUSTOM_FEES_FOR_TOKEN_WITH_FRACTIONAL_FEE, GET_CUSTOM_FEES_FOR_TOKEN_WITH_ROYALTY_FEE, HTS_GET_APPROVED, HTS_ALLOWANCE, HTS_IS_APPROVED_FOR_ALL will be fixed.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>
Signed-off-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>
Co-authored-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>