HIP-584 Historical: Add integration tests for custom fees and HTS calls - part 4 (#7476)
In this PR test cases for the previously failing custom fees tests and tests for HTS_GET_APPROVED, HTS_ALLOWANCE, HTS_IS_APPROVED_FOR_ALL are added. The latter are the only scenarios where an exception is not thrown when the data is not found for the passed historical block. The returned result is zeroes. The reason for this inconsistency is because the top frame in the stack overrides the output with zeroes and this behaviour comes from the Besu EVM. These cases should be improved in the future to behave like all the other scenarios where an exception is thrown.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>