HIP-584 Historical: Add integration tests for nested calls - part 3 (#7470)
In this PR three integration tests are added in order to help with the historical support testing. Each of them contains a function call which fails and a hardcoded value is returned after that. With this we aim to test if the expected hardcoded value is returned and that the failing function call does not make the whole transaction to fail. The tests are three in order to test the three main flows:
- get token info test - in this case we don't have a custom precompile in the code takes the shortcut for view functions
- HTS get approved test - we have a custom precompile in this case and the code does not take the shortcut for view functions
- mint nft token - a modification precompile. In this PR there is a fix in case an exception is thrown so that the frame would revert and the whole transaction would not fail.
---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>
Signed-off-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>
Co-authored-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>