Fix `NullPointerException` for `estimate_gas` when passing non-existent sender address (#7594)
Certain operations rely on the sender account of the call, and users have the flexibility to set this field to any valid address. However, if this specified account is not found in the database or in the state, the world updater will return null, leading to a `NullPointerException`.

In the Ethereum world, a valid gas estimation is returned even when passing an invalid or non-existent address.  This PR proposes a solution by creating a dummy account with the sender's address. This ensures that a valid gas estimation can be returned in such scenarios.

---------

Signed-off-by: Ivan Ivanov <ivanivanov.ii726@gmail.com>