Add freeze and unfreeze missing validations (#7973)
After changes, we don't have missing logic that validates the freeze and unfreeze state of the token.  I add `changeFrozenState` in `TokenRelationship` which modifies the state of the `frozen` property.  Before the property modification, the method performs validation, that the respective token has a freeze key. I replace `setFrozen` with `changeFrozenState` in `FreezeLogic`, `UnfreezeLogic` because this will help to properly change the state of the token. Make needed modifications in `TokenRelationshipTest`, `FreezeLogicTest`, `UnfreezeLogicTest`. The changes are borrowed from Hedera Services. 

---------

Signed-off-by: Zhivko Petkov <zhpetkov9@gmail.com>