Add pause and unpause missing validations (#7981)
After changes, we don't have missing logic that validates the pause state of the token. I add changePauseStatus in Token which modifies the state of the pause property. Before the property modification, the method performs validation, that the respective token has a pause key. I replace setPause with changePauseStatus in PauseLogic, and UnpauseLogic because this will help to properly change the state of the token. Make needed modifications in tests. The changes are borrowed from Hedera Services.

---------

Signed-off-by: Zhivko Petkov <zhpetkov9@gmail.com>