Fix address(this) resolution when mirror address is passed in the call (#7984)
There is an inconsistency on call address(this) function depending on the **to** parameter in a contract call request.
Address(this) will return whatever is specified as the 'to' parameter in the http request.
If a contract is created via ethereum transaction on hedera it will have a canonical ethereum address. It will also have the long zero address.
So in our contract call request we can put in the **to** parameter either the canonical evm address or the long zero and it should return the same response.
Currently if we call address(this) of a contract that has a canonical evm address but we pass it's long zero address in the contract call request it will return long zero address in the response. But we actually want the evm address returned.

This **PR** adds a check to see if a long zero address in passed in the request and will make the following verification:
If the address is long zero verify that there might be a canonical evm address.
If there exists a canonical evm address use it instead of the long zero.
Otherwise use the long zero.

Services uses similar approach -> 
```
if (relayerId == null) {
            result = evmTxProcessor.execute(
                    sender, receiver.canonicalAddress(), op.getGas(), op.getAmount(), callData, txnCtx.consensusTime());
}
```
 
This PR modifies:
`MirrorEvmTxProcessorImpl` adds a new method to determine the correct **to** address.
Added integration test for address function in `ContractCallDynamicCallsTest`

---------

Signed-off-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>