Add KYC key validations for grant and revoke logics (#8013)
After changes, we don't have missing logic that validates the grant and revokes kyc state of the token. I add changeKycState in TokenRelationship which modifies the state of the kycGranted property. Before the property modification, the method performs validation, that the respective token has a kyc key. I replace setKycGranted with changeKycState in GrantKycLogic, RevokeKycLogic, HederaTokenStore because this will help to properly change the state of the token. Make needed modifications in TokenRelationshipTest, HederaTokenStoreTest, GrantKycLogicTest, RevokeKycLogicTest, ContractCallServicePrecompileTest, ContractCallTestSetup. The changes are borrowed from Hedera Services.

---------

Signed-off-by: Zhivko Petkov <zhpetkov9@gmail.com>