Add a check so that prng and exchange rate precompiles won't accept value (#8206)
Adds check in Prng and Exchange rate precompiles so that they are not able to accept value. Seems like services added the change only for prng but not for exchange rate so they need to add one more pr there for exchange rate.

After https://github.com/hashgraph/hedera-services/issues/11218 was fixed in services, we need to port the changes to mirror node.

Current pitfall of web3 is that even though with the new checks we throw `INVALID_FEE_SUBMITTED` status. The end result will be a `CONTRACT_REVERT_EXECUTED`. Future work will be a work package to improve error handling logic and return proper statuses.

* Add new check to `ExchangeRatePrecompiledContract` to not accept value
* Add new check to `PrngSystemPrecompiledContract` to not accept value
* Modify `PrngSystemPrecompiledContractTest` to work with the new changes.
* Modify `ContractCallSystemPrecompileTest` to add new integration tests to check that prng and exchange rate do not accept value, added new functions in solidity files to test that, and made a small refactoring of the test class

---------

Signed-off-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>