Fix hollow account creation fail (#8200)
On transferFromNFT or transferFrom HTS call to a hollow account the transaction fails as the account could not be found in the store. This PR adds the missing implementation from services that creates an account in case it does not exist. Another missing implementation for the system account case was added - TransferPrecompile#revertIfReceiverIsSystemAccount. After these changes an error is thrown when a transfer is attempted to a system account. Some of the integration tests are adjusted as well - a new system account is defined in order to be used for the system account related tests and the old owner account is now used in the regular cases.

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>