Add gas throttle implementation (#8266)
Currently we are only using a rate limiter per request in the web3 module. But if there are a lot of requests that require a small amount of the gas the system may be a able to handle a larger number of request than in the current rate limit capacity.

This PR introduces a new limiter based on the used gas per second. It will continue to work alongside the existing rate limiter. The new limiter uses the bucket4j's `SynchronizedBucket`. The flow is like this:
1. When a request is made the passed `gasLimit` parameter value is consumed by the bucket.
2. After the request is processed the unused gas is returned back to the bucket and is available to be consumed by another request.

There are two new properties that are introduced:
- `hedera.mirror.web3.throttle.gasPerSecond` - the default value is 25 000 000. This is the amount of gas that can be consumed per second. The bucket automatically restores this amount as available to be used on each second.
- `hedera.mirror.web3.throttle.gasLimitRefundPercent` - the default value is 20. This is the maximum percent of the gas limit that will be returned back to the bucket if it has remained unused. For example, if the passed gas limit in the request is 1 000 000, the transaction used 200 000 gas, the unused gas is 800 000. The minimum of 800 000 and 20% of the initial 1 000 000 = 200 000 will be returned to the bucket, so in this case this will be 200 000 gas. This implementation is similar to the one in services. Without this percent there are cases such as an `EXCEPTIONAL_HALT` where the remaining gas in the frame is cleared by the EVM and we need to have a way to control how much gas needs to be restored.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>