Add gas limit calculation and lower default contract call gas (#8396)
The gas refund intends to force people to set an appropriate gas limit. In acceptance tests, I put a proper limit of 3M—the changes lower acceptance gas limit values. I add a new implementation of contractCallRequest in ModelBuilder and use actualGas in the gas limit calculation. The gas limit is 30% bigger than actualGasUsed when we call a contract in acceptance tests. When we don't have actualGasUsed we use the default value of the appropriate gas limit. The PR needs to make changes to support new gas limits based on actual gas used. The changes are in AbstractEstimateFeature, AbstractFeature, EstimateFeature, EstimatePrecompileFeature, HistoricalFeature, ModelBuilder.

---------

Signed-off-by: Zhivko Petkov <zhpetkov9@gmail.com>