Refactor ContractCallDynamicCallsTest (#8912)
With this PR the ContractCallDynamicCallsTest integration test is refactored using the web3j plugin and the big enum is now replaced with multiple junit tests so that it is easier to debug a particular scenario. This way also the number of persisted entities per test is reduced significantly and now we persist only the ones that are needed in the particular test. With this implementation the execution time of the ContractCallDynamicCallsTest is reduced 5 times (from about 50 seconds to 9-10 seconds, tested locally).

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>