Refactor `ContractCallServiceERCTokenTest` (#8985)
With this PR the ContractCallServiceERCTokenTest integration test file is partially refactored using the web3j plugin and the big enum is now replaced with multiple junit tests so that it is easier to debug a particular scenario. This way also the number of persisted entities per test is reduced significantly and now we persist only the ones that are needed in the particular test.

This PR refactors and extracts the read only erc contract functions in separate class, so it can be more readable. The modification functions will be extracted in another class with another PR

---------

Signed-off-by: Nikolay Nikolov <nikolay.nikolov@limechain.tech>
Signed-off-by: Nikolay Nikolov <96113577+nickeynikolovv@users.noreply.github.com>