Refactor ContractCallNestedCallsTest (#8955)
With this PR the ContractCallNestedCallsTest integration test is refactored using the web3j plugin and the big enum is now replaced with multiple junit tests so that it is easier to debug a particular scenario. This way also the number of persisted entities per test is reduced significantly and now we persist only the ones that are needed in the particular test. With this implementation the execution time of the ContractCallNestedCallsTest is reduced 3 times (from about 38 seconds to 13 seconds, tested locally).

The NestedEthCallContractFunctionsNegativeCases and NestedEthCallContractFunctions enums are moved from ContractCallNestedCallsTest to ContractDebugServiceTest and OpcodeServiceTest respectively as they still need these enums and these test classes are not refactored yet.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>