Return 501 NOT_SUPPORTED error response for unsupported precompiles (#9189)
This PR aims to protect the web3 calls to reach unimplemented precompiles due to missing functionality in the copied mono code base. When the json-rpc-relay tries to invoke such precompile it will receive the informative 501 error and perform the needed fallback logic. Currently they receive 500 INTERNAL_ERROR with CONTRACT_REVERT_EXECUTED status, which might be misleading.

The PR adds:

    logic to return NOT_SUPPORTED status for any new dynamic/modification precompile
    logic to return NOT_SUPPORTED status for HRC_IS_ASSOCIATED view precompile. That's the only read only precompile added in services after switching to modularized code and is not supported in hedera-mirror-web3. Unfortunately, we don't have a way to add a universal check for each not supported view precompile, since there is not distinction between not found precompile and failed input validation for the view precompiles, so we will have to add a check for each not supported precompile explicitly

---------

Signed-off-by: IvanKavaldzhiev <ivankavaldzhiev@gmail.com>