Add check for `web3j-cli` existence before installation (#9224)
As investigated by @kselveliev, it seems that the web3j-cli download script has a rate limit. This PR adds a check if it is already installed. A similar check exists in the download script but this PR aims to fix the rate limit issue by performing the check earlier - in our gradle task.

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>