Update contracts to latest versions from the smart contracts repo (#9279)
This PR sets the licenses of the used test contracts wherever they are currently missing. This aims to reduce the number of warnings during build. Also, the HederaTokenService and IHederaTokenService are updated to their latest versions from the smart contracts repo. The changes are mostly in the comments there. Nothing significant.

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>