Split the `ContractCallServicePrecompileTest` into two classes (#9277)
This PR splits the ContractCallServicePrecompileTest into two classes based on the readonly tests and the modification tests. This change is needed for a followup PR that aims to remove the duplicated methods in classes. As a result each of the two classes uses only one contract. There is no change in the actual code of the tests.

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>