Remove `web3j-cli` (#9303)
This PR removes the custom scripts that download the web3j-cli and compile the historical contracts with it. The need for that is that the scripts are not fully stable. Also, the download script always downloads the latest web3j-cli version (if the cached local folder is removed), so this way the latest version could possibly contain some breaking changes that might break our code as well. Now the solidity-gradle-plugin and the web3j-gradle-plugin are used as gradle dependencies. This fixes the issues that are described above.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>