Fix error on calls with value with not existing sender alias (#9302)
This PR aims to fix null error when a contract call is made with value and evm address as a sender. Where the evm address is not existing.

Couple of scenarios:

Currently when an empty from is passed. 
We are creating an Account.getDummySenderAccount(). that does not have alias

If the from is a not existing long zero address
We are creating an Account.getDummySenderAccount(long zero address) from it without alias.

**In the cases where evm address alias was passed as a from parameter. We were still creating a dummy account without alias. And subsequent account retrievals by alias were returning null causing the error since we did not create the account with that alias.**

**This pr adds a check.
Check if the from is an evm canonical address. Create the dummy account with that alias.
Otherwise create the dummy account as before without alias.**



This PR modifies:
`MirrorEvmTxProcessorImpl` - add a check to see is the from address is not existing and also a canonical evm address and it it is creates the dummy account with that alias.

`Account`.java - created new methods that creates dummy account with and alias

`MirrorEvmTxProcessorTest` - adds a test case for the error scenario.

`ContractCallServiceTest` - adds a test case for the error scenario.

---------

Signed-off-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>